# optimize the backend calls we make for frame creation