--- a/dwds/lib/src/debugging/instance.dart
+++ b/dwds/lib/src/debugging/instance.dart
@@ -84,6 +84,9 @@ class InstanceHelper extends Domain {
     if (primitive != null) {
       return primitive;
     }
+
+    // TODO: This is checking the JS object ID for the dart pattern we use for
+    // VM objects, which seems wrong (and, we catch 'string' types above).
     if (isStringId(remoteObject.objectId)) {
       return _stringInstanceFor(remoteObject, offset, count);
     }
@@ -94,6 +97,7 @@ class InstanceHelper extends Domain {
     if (metaData.jsName == 'Function') {
       return _closureInstanceFor(remoteObject);
     }
+
     var properties = await inspector.debugger.getProperties(
         remoteObject.objectId,
         offset: offset,
@@ -319,6 +323,7 @@ class InstanceHelper extends Domain {
     if (remoteObject == null) {
       return _primitiveInstanceRef(InstanceKind.kNull, remoteObject);
     }
+
     switch (remoteObject.type) {
       case 'string':
         var stringValue = remoteObject.value as String;