Optimize the backend calls we make for frame creation:
- rev to use the latest version of `package:webkit_inspection_protocol` (note: as package:test depends on webkit_inspection_protocol, this will select an older version of test until package:test is republished)
- use the `this` object from the JS frame instead of making a remote eval call
- don't iterate over the (very large) DDC closure scopes in JS frames
- add the synthetic 'return value' variable to a returned frame if its available (this isn't done by the VM implementation, but will be useful to add to the VM)
- we now display local variables in the order they appear in the method (instead of sorted lexically); this maps well to how other debuggers work
- fix an issue where the dart frame index wouldn't align with the js frame we were using to do things like eval (they would get out of alignment if we didn't create a dart frame for a JS frame on the call stack)
- when encountering an exception while computing the stack frames, we still send the paused event to the client (w/o frame info)

### Times
For a reasonably sized flutter web app (flutter/gallery), creating frames when paused at a breakpoint or stepping took around 23s-24s (23,299ms). Flutter apps typically have deep call stacks - for this case, the call stack was 170-180 frames, and each frame took around 140ms to populate.

After this change, the total frame population time is around 700-800ms; stepping takes 2-3s, with perhaps 1,500ms taken by the JS VM to resume execution, hit the next breakpoint, and send us the JS frame info. DDC changes to create smaller closures may help here? Other improvements we can make on the package:dwds side are listed below.

### Some possible improvements in the future:
- caching classrefs - this is likely the single biggest remaining win. We call a backend eval call ~per JS property to populate the ClassRef info for that BoundVariable. These computed ClassRefs could be cached per JS class name, which would save us many backend calls per frame
- we call many backend eval calls when populating a frame; we should look as creating a single optimized JS eval in order to populate the frame info in one backend call
- investigate the use of `package:pool` to create ~6 frames at a time instead of each frame sequentially
- the JS frame now includes info about async frames; we could choose to use this to display additional dart frames. This could help users puzzle out the logical call chain for async calls