@devoncarew 2020-05-20
I'm seeing:

```
pub run test -x frontend-server
00:00 +0: test\build_daemon_breakpoint_test.dart: shared context (setUpAll)
00:58 +0 -1: test\build_daemon_breakpoint_test.dart: shared context (setUpAll) [E]
  SessionNotCreatedException (33): session not created: This version of ChromeDriver only supports Chrome version 81
    (Driver info: chromedriver=81.0.4044.138 (8c6c7ba89cc9453625af54f11fd83179e23450fa-refs/branch-heads/4044@{#999}),platform=Windows NT 10.0.17763 x86_64)
  package:webdriver/src/handler/json_wire/utils.dart 79:9     parseJsonWireResponse
  package:webdriver/src/handler/json_wire/session.dart 26:21  JsonWireSessionHandler.parseInfoResponse
  package:webdriver/src/handler/json_wire/session.dart 18:7   JsonWireSessionHandler.parseCreateResponse
  package:webdriver/src/common/request_client.dart 96:25      AsyncRequestClient.send
```

on the windows `pub run test` bot; not sure it's related to this PR.