@grouma 2020-05-20
> Looks like the same CI failure is hitting #1018 as well.

We use [chocolatey](https://github.com/dart-lang/webdev/blob/fddcaa39347085bb5698a26c039cf2534f10e9a6/tool/travis_setup.sh#L7) to install Chrome and Chrome Diver. It looks like just one of the repos were updated. It'll probably resolve itself once both are updated. It's weird that they decided to update them out of sync.